refactor: moved api error hand test to main test
